ZDM #110.

+++ kernel/include/linux/blkzoned_api.h
@@ -16,9 +16,9 @@
 
 #include <uapi/linux/blkzoned_api.h>
 
-extern int blkdev_issue_zone_action(struct block_device *, unsigned long bi_rw,
+extern int blkdev_issue_zone_action(struct block_device *, unsigned int bi_rw,
 				    sector_t, gfp_t);
-extern int blkdev_issue_zone_report(struct block_device *, unsigned long bi_rw,
+extern int blkdev_issue_zone_report(struct block_device *, unsigned int bi_rw,
 				    sector_t, u8 opt, struct page *, size_t,
 				    gfp_t);
